+++ src/px-script/main/action-list/index.test.js
@@ -9,9 +9,9 @@ describe("px-script: action-list", () => {
     document.body.appendChild(div);
 
     const ActionList = ({ active }) => (
-        <div className="action-list">
+        <div className={`action-list${active ? " active" : ""}`}>
             <i className="material-icons">more_vert</i>
-            <div className={`action-menu${active ? " active" : ""}`}>
+            <div className="action-menu">
                 <a href="#"><i className="material-icons">bookmark</i>Add bookmark</a>
                 <a href="#"><i className="material-icons">business_center</i>Add client</a>
             </div>
@@ -32,17 +32,15 @@ describe("px-script: action-list", () => {
 
         const renderedActionList = document.querySelector(".action-list");
         const toggleBtn = renderedActionList.querySelector("i.material-icons");
-        const actionMenu = renderedActionList.querySelector(".action-menu");
 
         expect(renderedActionList).toBeDefined();
         expect(toggleBtn).toBeDefined();
-        expect(actionMenu).toBeDefined();
-        expect(actionMenu.classList).not.toContain("active");
+        expect(renderedActionList.classList).not.toContain("active");
 
         actionList.init();
         toggleBtn.dispatchEvent(new Event("click"));
 
-        expect(actionMenu.classList).toContain("active");
+        expect(renderedActionList.classList).toContain("active");
 
         ReactDOM.unmountComponentAtNode(div);
     });
@@ -52,17 +50,15 @@ describe("px-script: action-list", () => {
 
         const renderedActionList = document.querySelector(".action-list");
         const toggleBtn = renderedActionList.querySelector("i.material-icons");
-        const actionMenu = renderedActionList.querySelector(".action-menu");
 
         expect(renderedActionList).toBeDefined();
         expect(toggleBtn).toBeDefined();
-        expect(actionMenu).toBeDefined();
-        expect(actionMenu.classList).toContain("active");
+        expect(renderedActionList.classList).toContain("active");
 
         actionList.init();
         toggleBtn.dispatchEvent(new Event("click"));
 
-        expect(actionMenu.classList).not.toContain("active");
+        expect(renderedActionList.classList).not.toContain("active");
 
         ReactDOM.unmountComponentAtNode(div);
     });
@@ -71,16 +67,14 @@ describe("px-script: action-list", () => {
         ReactDOM.render(<ActionList active />, div);
 
         const renderedActionList = document.querySelector(".action-list");
-        const actionMenu = renderedActionList.querySelector(".action-menu");
 
         expect(renderedActionList).toBeDefined();
-        expect(actionMenu).toBeDefined();
-        expect(actionMenu.classList).toContain("active");
+        expect(renderedActionList.classList).toContain("active");
 
         actionList.init();
-        actionMenu.querySelector("a").click();
+        renderedActionList.querySelector("a").click();
 
-        expect(actionMenu.classList).not.toContain("active");
+        expect(renderedActionList.classList).not.toContain("active");
 
         ReactDOM.unmountComponentAtNode(div);
     });
@@ -89,16 +83,14 @@ describe("px-script: action-list", () => {
         ReactDOM.render(<ActionList active />, div);
 
         const renderedActionList = document.querySelector(".action-list");
-        const actionMenu = renderedActionList.querySelector(".action-menu");
 
         expect(renderedActionList).toBeDefined();
-        expect(actionMenu).toBeDefined();
-        expect(actionMenu.classList).toContain("active");
+        expect(renderedActionList.classList).toContain("active");
 
         actionList.init();
         document.querySelector("html").click(); // for clicking outside action list
 
-        expect(actionMenu.classList).not.toContain("active");
+        expect(renderedActionList.classList).not.toContain("active");
 
         ReactDOM.unmountComponentAtNode(div);
     });