+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## [2.0.2] - 2019-03-07
+## [2.0.3] - 2019-03-25
 
 ### Fixed
 
-- Fixed issue with "has-vscroll" class not being removed on sheet close.
-- Fixed issue with panel table css selectors.
+- Fixed issue with action-list icon showing through action-list menu.
+- Fixed issue with polyfills